feat! : passkeys change implemented

+++ app/backend/jwt/index.ts
@@ -8,6 +8,10 @@ export async function checkUser(){
     .select()
     .eq("secret",sessionStorage.getItem("sessionSecret")!)
 
+    if (sessionCheck.data![0] == null){
+      return;
+    }
+
     if (sessionCheck.data != null && !((jwt_decode(sessionCheck.data[0].loginSessionToken) as any).userName == sessionStorage.getItem("user")!)){
       alert("Invalid Authentication!")
       sessionStorage.clear()
